Refactor video tile generation code and fix text positioning issue
This commit refactors the code for generating a tile of the video frames and fixes an issue with the positioning of the text on the tile. Previously, the scale parameter was placed after the tile parameter, which caused the padding and margin values to be calculated incorrectly. This commit fixes the issue by calculating the tile_img_width separately and using it to set the scale parameter before the tile parameter.

In addition, the commit updates the ffmpeg_cmd used to generate the final composite image by setting the x parameter of the drawtext filter to $margin instead of $margin/2, which correctly centers the text on the tile.

The commit also cleans up some commented-out code and adds logging to an output file.